+++ v8go.h
@@ -5,6 +5,7 @@ extern "C" {
 #endif
 
 #include <stddef.h>
+#include <stdint.h>
 
 typedef void* IsolatePtr;
 typedef void* ContextPtr;
@@ -49,6 +50,13 @@ extern RtnValue RunScript(ContextPtr ctx_ptr,
 
 extern void ValueDispose(ValuePtr ptr);
 const char* ValueToString(ValuePtr ptr);
+const uint32_t* ValueToArrayIndex(ValuePtr ptr);
+int ValueToBoolean(ValuePtr ptr);
+int32_t ValueToInt32(ValuePtr ptr);
+int64_t ValueToInteger(ValuePtr ptr);
+double ValueToNumber(ValuePtr ptr);
+const char* ValueToDetailString(ValuePtr ptr);
+uint32_t ValueToUint32(ValuePtr ptr);
 int ValueIsUndefined(ValuePtr ptr);
 int ValueIsNull(ValuePtr ptr);
 int ValueIsNullOrUndefined(ValuePtr ptr);